+++ packages/docz-tools/CHANGELOG.md
@@ -8,6 +8,30 @@ menu: Changelog
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.8.7](https://github.com/GetJobber/atlantis/compare/@jobber/docz-tools@1.8.6...@jobber/docz-tools@1.8.7) (2021-09-23)
+
+**Note:** Version bump only for package @jobber/docz-tools
+
+
+
+
+
+## [1.8.6](https://github.com/GetJobber/atlantis/compare/@jobber/docz-tools@1.8.5...@jobber/docz-tools@1.8.6) (2021-09-23)
+
+**Note:** Version bump only for package @jobber/docz-tools
+
+
+
+
+
+## [1.8.5](https://github.com/GetJobber/atlantis/compare/@jobber/docz-tools@1.8.4...@jobber/docz-tools@1.8.5) (2021-09-22)
+
+**Note:** Version bump only for package @jobber/docz-tools
+
+
+
+
+
 ## [1.8.4](https://github.com/GetJobber/atlantis/compare/@jobber/docz-tools@1.8.3...@jobber/docz-tools@1.8.4) (2021-09-22)
 
 **Note:** Version bump only for package @jobber/docz-tools